added disparity_view/util.py
