+++ disparity_view/util.py
@@ -1,0 +1,15 @@
+import numpy as np
+
+
+def dummy_camera_matrix(image_shape) -> np.ndarray:
+    # 近似値
+    cx = image_shape[1] / 2.0
+    cy = image_shape[0] / 2.0
+
+    # ダミー
+    fx = 1070  # [pixel]
+    fy = fx
+
+    # カメラパラメータの設定
+    camera_matrix = np.array([[fx, 0, cx], [0, fy, cy], [0, 0, 1]])
+    return camera_matrix
