+++ scripts/gen_ply.py
@@ -4,19 +4,7 @@ import cv2
 import numpy as np
 import open3d as o3d
 
-
-def dummy_camera_matrix(image_shape) -> np.ndarray:
-    # 近似値
-    cx = image_shape[1] / 2.0
-    cy = image_shape[0] / 2.0
-
-    # ダミー
-    fx = 1070  # [pixel]
-    fy = fx
-
-    # カメラパラメータの設定
-    camera_matrix = np.array([[fx, 0, cx], [0, fy, cy], [0, 0, 1]])
-    return camera_matrix
+from disparity_view.util import dummy_camera_matrix
 
 
 def gen_ply(disparity: np.ndarray, left_image: np.ndarray, outdir: Path, left_name: Path, baseline=100.0 ):
