+++ scripts/reproject.py
@@ -11,21 +11,7 @@ import cv2
 from tqdm import tqdm
 
 from disparity_view.reprojection import reproject_from_left_and_disparity, generate_point_cloud, reproject_point_cloud
-
-
-def dummy_camera_matrix(image_shape) -> np.ndarray:
-    # 近似値
-    cx = image_shape[1] / 2.0
-    cy = image_shape[0] / 2.0
-
-    # ダミー
-    fx = 1070  # [pixel]
-    fy = fx
-
-    # カメラパラメータの設定
-    camera_matrix = np.array([[fx, 0, cx], [0, fy, cy], [0, 0, 1]])
-    return camera_matrix
-
+from disparity_view.util import dummy_camera_matrix
 
 def gen_right_image(disparity: np.ndarray, left_image: np.ndarray, outdir: Path, left_name: Path):
     """
